My i3 config.
This is my i3 config. It's nothing special but I'm keeping it for progress so just in case if I have to redo something, I have an easy restore point. I always keep a few backups but if anybody wants to feel free to change anything or slim it down you can if you want to. I wouldn't mind.